requirements.txt
torch>=2.2.0,<2.4.0
torchvision>=0.17.0,<0.19.0
scikit-learn>=1.4.0
scikit-image>=0.22.0
albumentations>=1.4.0
opencv-python>=4.8.0
pillow>=10.2.0
pandas>=2.2.0
matplotlib>=3.8.0
seaborn>=0.13.0
tqdm>=4.66.0
pyyaml>=6.0.0
numpy>=1.26.0
